requirements.txt
openai
pyyaml
rich
simplejson
